effectors: split DisplayEffector into BrightnessEffector and DPMSEffector
