+++ src/control/effector_inventory.rs
@@ -13,12 +13,13 @@ use anyhow::Result;
 use std::collections::HashMap;
 
 pub fn get_known_effector_names() -> Vec<&'static str> {
-    vec!["display", "session", "sleep", "lock"]
+    vec!["brightness", "dpms", "session", "sleep", "lock"]
 }
 
 pub fn get_effects_for_effector(effector_name: &str) -> Vec<Effect> {
     match effector_name {
-        "display" => system::display_effector::DisplayEffector.get_effects(),
+        "brightness" => system::brightness_effector::BrightnessEffector.get_effects(),
+        "dpms" => system::dpms_effector::DPMSEffector.get_effects(),
         "session" => system::session_effector::SessionEffector.get_effects(),
         "sleep" => system::sleep_effector::SleepEffector.get_effects(),
         "lock" => system::lock_effector::LockEffector.get_effects(),
@@ -33,8 +34,13 @@ pub async fn spawn_effector<B: BrightnessController, D: DisplayServer>(
 ) -> Result<EffectorPort> {
     let config_clone = config.map(|c| c.clone());
     match effector_name {
-        "display" => {
-            system::display_effector::DisplayEffector
+        "brightness" => {
+            system::brightness_effector::BrightnessEffector
+                .spawn(config_clone, dependency_provider)
+                .await
+        }
+        "dpms" => {
+            system::dpms_effector::DPMSEffector
                 .spawn(config_clone, dependency_provider)
                 .await
         }
