+++ src/system/mod.rs
@@ -1,4 +1,5 @@
-pub mod display_effector;
+pub mod brightness_effector;
+pub mod dpms_effector;
 pub mod inhibition_sensor;
 pub mod lock_effector;
 pub mod session_effector;
