+++ src/system/test/mod.rs
@@ -1,4 +1,5 @@
-mod display_effector_test;
+mod brightness_effector_test;
+mod dpms_effector_test;
 mod inhibition_sensor_test;
 mod lock_effector_test;
 mod session_effector_test;
